MB-24655 system indexers don't read from stop channel
This is a partial fix - as we have to implement a way to notify
foreach / remote routines of stop signals to complete the fix.

Essentially, since day one system indexers have ignored stop channels,
which would lead to leaked go routines, and locks left behind.
This can be one of the causes for MB-24215 - although, from the
diagnostics I have seen, there are others.

Change-Id: I982061dbbe769ce9bfaa68f2169d9d9dd2e8defc
Reviewed-on: http://review.couchbase.org/78971
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>