MB-24215 requests don't move to completed MB-24655 system scans don't monitor the stop channel
A few changes here:
- sendSystemKey has a timeout. this will need to be sync'd with the fetch
timeout

- the 3 foreach functions now have a nonBlocking and a blocking function
the idea is that we set up something in the non blocking part, release
the lock, and if required do blocking work with the second functions

- the arguments to the 3 foreach functions and GetRemoteKeys() now have
to return a boolean to indicate whether to continue the foreach

- genCache has been amended to offer nonblocking and blocking
operation

The system keyspaces and caches have been amended to adapt to the above.
To make things complete for active_requests, we have to code MB-24656
(move active requests to genCache).

Change-Id: I852ade97c0e0af7191554c6df5c25c4b11b2c94e
Reviewed-on: http://review.couchbase.org/79065
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>