MB-24677 use partial MRU sorting
Last change - I promise.
- use a larger number of cache buckets. it's not as if they are going
to use lots of memory, so we might as well enjoy the higher number of
locks.

- use partial MRU sorting.
entries are given an MRU sorting timestamp, and there's a global MRU
timestamp tracked every time an entry is brought to the top of the cache.
the idea is that for entries that have been used very recently, we don't
care that their position is accurate, because they are not at risk of
being cleaned.
so if they are in the top half of the cache, we can just avoid bringing
them to the top, which saves time.
If they are in the bottom half, we don't want them slipping away, so we
will promote them.

For those curious, and in the know - this is akin to RSAM buffer manager
implemented in 9.4

- use shared locks for MRU tracking where possible.
if the Use() method is passed a nil process function (nothing to do) and
the cache is nowhere near being cleaned (ie less than half full), we can
change the Use() to a Get() and use shared locks instead!

Change-Id: I8e6e8a104cb31ebc7a3d5f597a6c499fd5edc7a1
Reviewed-on: http://review.couchbase.org/79187
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>