MB-25317 joins turn off covering scans in error
The underlying issue here is that CoveredBy has always used boolean
return values, with true meaning anything from yes, this is covered, to
don't know yet, keep on going, to this keyspace is unrelated, skip this
expression list.

Changed CoveredBy() to use 4 state results, so that appropriate actions
can be taken such as skip or continue.
ExpressionBase.CoveredBy() will skip a whole expression list if
Identifier.CoveredBy reports a different keyspace.

For now, we return CoveredTrue in most places, but as issues arise, we
can change the actual outcome as needed.

Added a wrapper for use by the planner.
Added more tests.
the gsi tests are run one at a time because cbauth would get stale
on subsequent tests.

Change-Id: Iefb940c7e13952df413d22b09241a0608cfda27c
Reviewed-on: http://review.couchbase.org/81237
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>