MB-25370 covering code covers variables in error
The underlying issue is that Identifier.CoveredBy() does not differentiate
in between identifiers type, so covers collection variables as well as
keyspaces.
Even though there's support to mark expressions as variables, that is not
currently used anywhere, and variables are not identified as such.
Ideally we should augment bindings not to use strings for variable names
but expressions, and then use SetCollectionVariable in the parser and in 
the Copy operator to mark them as such.
That's an awful lot of code.
An other alternative is to mark them via the collection Children() methods,
but I haven't yet found an easy way to do that.

I have started infrastructure for the above, but for the time being,
nav_field.CoveredBy() will not cover identifier that do not match the
keyspace by propagating the skip operation.

I have switched the CoveredBy() 'single' parameter to an option structure
so that as we add behavioral changes, we don't have to change interface.

Change-Id: I62cd1812c22cd6005345e3d33072bd3452d7bc09
Reviewed-on: http://review.couchbase.org/81273
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>