MB-25494 panic in centos
I don't have a reproducible testcase, except by instrumenting code and
adding delays - but the theory, put forward by Sitaram, is plausible.

In general terms, on very loaded systems it is possible that requests
time out before RunOnce() has had a chance to run, and therefore by the
time they run, the resources have been cleaned, because they hadn't been
claimed in the first place.

The fix is to do away with the waitGroup, and instead have a condition,
with two states for the operator: primed and completed.

When operators enter base.active(), this will check (under lock) if the
operator has already been completed - if it has, RunOnce() will just
terminate, if not the operator will be marked as primed.
base.inactive() will mark the operator as completed and wake up the
freeer, if there.
base.wait() will wait for execution termination on the condition, if 
necessary.

I have tried this under load and I don't see any ill effect.
Locking is only done 3 times in the lifetime of the operator, and the
general case is that the lock will be found unlocked (let alone a need
for waiting).

I have also standardised a couple of logging.Stackf in the same area
to asserts, to make the code a bit more maintainable.

Change-Id: I33e04b256e511520a24bc60780050af7e31c6b20
Reviewed-on: http://review.couchbase.org/81844
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>