MB-25949
This is a regression from MB-25756 (Improve expression handling in planner).

The fix for MB-25756 separated the WHERE clause expression into filters belonging to each keyspace.
Then during index selection only expression for the appropriate keyspace is used. The issue with
this bug is for a query in the form:

   FROM default a UNNEST array as b WHERE b.field > 1 and b.field < 5

in this case the filters on b.field belongs to keyspace 'b', however, since 'b' is just an element
of an array in 'a', during index selection for 'a' we should try to match the 'b' filters with
array index definition on 'a'. This was done previously since the entire WHERE clause is used,
however with the fix for MB-25756 the 'b' filters are now missing when considering unnest scan with
array indexes (buildUnnestScan()).

The fix is to add all filters from keyspace 'b' to consideration when UNNEST is being done so array
index on 'a' can be matched.

Change-Id: Id7fbc6468b8ca3f6a9881f2453d2214e1328f657
Reviewed-on: http://review.couchbase.org/83163
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>