MB-25957
This is a regression from MB-25756 (Improve expression handling in planner).

The planner now expects baseKeyspaces map to be setup in the builder before
building scans, etc. For select statements this is setup by traversing the
from clause, and for delete/update statements this is done in beginMutate()
function. For merge statements this is currently not done, resulting in error.

The fix is to setup builder.baseKeyspaces map in VisitMerge().

Change-Id: Ida4927424124df4979f2582c3a37c8ab2c513d4b
Reviewed-on: http://review.couchbase.org/83175
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>