MB-26205: Incorrect index span setting after fix for MB-26132
In patch set 2 for fix of MB-26132 new code is added to skip OR predicate
for index span setting. This causes issues when we actually have multiple
OR predicate ANDed together (and thus not caught by earlier check for OR).
Note that to see OR predicate at this part of code we must have passed
earlier checks for sargable predicates for index keys, which means that
each subterm of the OR must be sargable for the same (set of) index key(s),
therefore logic is added to handle OR predicate, which could potentially
be sargable for multiple index keys. (changes in sarg.go).

In addition, patch set 3 for fix of MB-26132 incorrectly assumed that
when copying baseKeyspaces only the current one is interesting. This will
cause issue when ANSI JOIN is supported since by doing so we'll lose
information on whether a filter is a join filter to start with. Thus
changes in patch set 3 are also reverted. (changes in rest of files).

Change-Id: I0071b23a2dae76854b0032c49742eae38f58ef10
Reviewed-on: http://review.couchbase.org/83841
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>