MB-25904 skip parsed value validation if coming from kv
This is part 1 of a 3 part fix: if the KV says that the document type is
JSON, we can skip validation entirely.
Sadly, memcached only reports the type if we have negotiated HELLO::JSON
which, as far as I have debugged, gocouchbase does not.
So the second part of the fix is to have gocouchbase go through HELLO.
The third part of the fix is to have the KV actually report the type,
which as I understand it, is currently disabled.

Change-Id: I330707f1845fc9f9c0ff33a8c0394a1b77a633a7
Reviewed-on: http://review.couchbase.org/83915
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>