MB-26489 execution pooling
This change is not execution pooling per se but, it does ease off the
load on the GC.

Cuurently when we populate an execution base, we 
- create a new object
- populate the new object
- copy the newly created base onto the Operator base
- finally ditch the newly created base.

This we do about 8 to 10 times for each request, which, for requests that
only yield one document, can be quite costly.

I haven't managed to set up clients yet to measure performance throughput
improvements properly yet, but I do know that for Q2 thi change
improves average latency by 20% (from 4800ns to 4000).

We have the same issue in Algebra (not a major thing) and in
expression (which might be major)


Change-Id: Ibe29658cbadf5409f4e8f420b6a11678354b8223
Reviewed-on: http://review.couchbase.org/84804
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>