MB-25757 Fix an issue in ANSI JOIN phase I support
When the planner is done planning for one keyspace, it goes through join
filters of other keyspaces and remove the reference to this keyspace
(whose planning is done), and if there is only a single keyspace
reference left, it'll move the filter from joinfilters to filters such
that this filter can potentially be used for index selection.

The issue is after we moved a join filter, we need to compact the
joinfilters slice, and the problematic situation is when joinfilters
started out with 2 entries, but both of them are set to nil (moved to
filters), then current code simply replace the first nil with the last nil,
resulting in a slice with a single nil entry which then causes problems
later on.

The fix is to first trim all nil entries from the end of the slice before
(and during) compacting process so we don't end up replacing a nil entry
with another nil entry.

Change-Id: I6703a8d0ea5c326fd3d29bafee7e1e184d614fef
Reviewed-on: http://review.couchbase.org/85408
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>