MB-25757 Add support for arrays in ANSI JOIN
Remove restrictions for array index keys in previous checkins, and enhance
SargForFilters() function to handle array index keys. The logic for handling
array index keys exist previously in function visitAndArrayKey().

Also modify logic for performing Cover() expression transformation to
transform ON-clause earlier, in case the entire ON-clause is transformed,
such that when plan.AnsiJoin or plan.AnsiNest is built we already have
the transformed ON-clause.

The logic for handling IN-clause (for cases where left-hand-side of join is
an array) already exists and should work as is for ANSI JOIN.

Change-Id: Ib3c60326daae78f72a9f3f26ad2de6d598bf2729
Reviewed-on: http://review.couchbase.org/85817
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>