MB-25757 Add ability to push ON-clause filters from inner joins to left-hand-side keyspace
The ON-clause of ANSI JOIN can contain arbitrary expressions, which means
it can contain filters for the left-hand-side keyspace. Currently the
ON-clause filters are only considered when right-hand-side keyspace is
being planned, thus when left-hand-side keyspace is being planned the
ON-clause filters are not considered.

We now "push" ON-clause filters from all inner joins to the left-hand-side
keyspace, and can be used for index selection for the left-hand-side
keyspace. This "push" is achieved by collecting ON-clauses from all inner
joins when the algebra tree is traversed and keyspace references are
collected. The combined ON-clauses can then be used for index selection
for the left-hand-side keyspace. As a side result, the function
combineFilters() require a "true" argument to include ON-clause filters
in serveral places when left-hand-side keyspace is being planned.

Change-Id: I343b20fe761190b9445c757923764b155ecd9f70
Reviewed-on: http://review.couchbase.org/85867
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>