MB-26946 incorrect explain plan for query with OR predicates
This is a regression from ANSI JOIN support phase I (MB-25757).

Previous checkin incorrectly assumes that if an OR predicate is sargable then
each arm of the OR must be sargable for the same index keys. In case of the
problematic query, the index has two keys (k1, k2), and the query predicate
looks like k1 = "windows" OR k2 > 1. We also derive a "k1 is not null"
predicate since the query has an aggregate on k1. When determining sargable
index keys the DNF form of the predicate (k1 = "windows" and k1 is not null)
or (k2 > 1 and k1 is not null) is used, and we concluded the predicate is
sargable with min sargLength of 1. However, with the new way of generating
sarg keys, individual filters are used, so we first try the OR predicate
(k1 = "windows" or k2 > 1) which is NOT sargable since the second arm of the
OR clause does not contain a predicate on k1.

The fix is to restore previous code where in case of OR predicate we use
SargFor() function (which takes the entire filter) instead of SargForFilters()
function (which uses each individual filter). For this to work properly
SargFor() and associated sargForOr() function now takes an additional argument
to indicate whether the filter is a join filter. Note since we are not using
individual filters in these functions we can only approximate and say if any
of the expression inside the OR clause involves join then we consider the
entire OR clause as join filter.

Also several functions in sarg.go takes a "total" argument (total length of
index keys) but it's not used anywhere so the "total" argument is removed from
several functions.

Change-Id: I440541c578368e982fecaef6328aa288a23482ac
Reviewed-on: http://review.couchbase.org/86092
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>