MB-26953 merge misses updates MB-27061 scans don't report service time
The underlying issue here is that merge.go:wrapChild() sets the insert / delete /update parent to the merge, but also as the operator to stop, so essentially the merge operator is being stopped twice: once normally via the child channel and also (incorrectly) via notifyStop.
The new value exchange always gives precedence to stop, this was not necessarily the case with the old one, hence the one change to be applied never made it past the set operator.
The only stop mechanism to be used for merge is the child channel. 

Change-Id: Ia3a9281732bdd5282dec08dfcc513d93dde99595
Reviewed-on: http://review.couchbase.org/86271
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>