MB-27086 keepalive operators don't close on exit
I run into this when I mistakenly run the tests with no buckets, and multistore hanged.
The only important one is authorise.go:RunOnce() on the authentication failure.
Since close isn't done, Done() will wait forever for the operator to complete.
Everything else is not important, just there for completeness.

Change-Id: Idc01ef351d8dbd23bd480889e6ef0c26e77b2546
Reviewed-on: http://review.couchbase.org/86319
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>