MB-27109 Unexpected error in ANSI JOIN if left-hand-side of join is a subquery
This is a regression from the change for pushable ON-clause filters for ANSI
JOIN (MB-25757). For that change we first gather all ON-clause filters from
inner joins in the query and process the gathered ON-clause filters as we
process the WHERE clause filters - when we perform index selection for the
first keyspace in the query. In case of the defect, the left-hand-side of
the ANSI JOIN is not a keyspace (it's a subquery), as a result, we did not
go through the code that would process the WHERE clause and the pushable
ON-clause filters, leading to issues in the planner.

Previously the WHERE clause is processed as we process the first keyspace
since that's the only keyspace that the planner does index selection for.
Now since index selection could be done for all keyspaces in an ANSI JOIN
query, we move the logic for processing WHERE and pushable ON-clause filters
to the top level where the FROM clause is being processed, such that this
process will happen regardless of whether the left-hand-side of a join is
a keyspace or not.

Change-Id: Ie5f8bdf4b85d7e947e37c1b0ff945f5271505faf
Reviewed-on: http://review.couchbase.org/86407
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>