MB-27138 Expression not properly classified if whole keyspace is used
As part of expression handling work (MB-25756) a filter is classified as
belonging to a keyspace if it references the keyspace, however, it
incorrectly assumes the filter is always in the form of keyspace.field.
In this defect, we have a UNNEST with an alias v1, and the predicate
just references v1 directly (v1 = 10), not a field of v1, and thus
current code does not properly classify the filter.

Fix is to remove the logic for checking field reference. Since we
compare an identifier with existing keyspace names (gathered
previously) this should not add too much overhead.

Change-Id: Icac848b1bab52efae3d771c733093a06ea29974a
Reviewed-on: http://review.couchbase.org/86491
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>