MB-25757 Add ability to derive IS NOT NULL predicate based on join filter
In ANSI JOIN if there is no filter on the first keyspace then currently
only primary index scan is possible. However, if there is a join predicate
in the form of b1.c1 = b2.c2 which is in either the WHERE clause or the
ON clause of an inner join, then we should be able to derive a predicate
in the form of "b1.c1 IS NOT NULL", in case there exists an index with
b1.c1 as the leading index key, such a predicate can then be used to
generate an index span, thus allowing a secondary index scan.

This checkin adds such a capability. It first gathers leading index key
expression on all secondary indexes, then looks through existing filters
to exclude the keys that already has a filter. Next it looks through all
join filters looking for references to any of the remaining index key
expressions. If found, it means a derived IS NOT NULL filter can then
be used to generate an index span, in this case the IS NOT NULL filter
is derived.

Note currently only the leading index key is being considered, since
an index scan requires a filter on the leading index key. Filters on
other index keys can be derived, however that adds more complexity
(only useful when a filter exists on all previous index keys) and
thus currently it is not done.

The logic for generating covering primary index scan is also modified
such that after potential filter derivation, only when there is still
no filter on the keyspace do we attempt to do covering primary scan.
Note covering primary scan is still done before regular primary scan.

Change-Id: I95264aaf805de20494366d079cb63751f63f6e54
Reviewed-on: http://review.couchbase.org/86533
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>