MB-27148 Engine panic during Index Join if left-hand-side is not using covered index scan
During index join execution, if the right-hand-side of index-join is using covered index
scan, we call joinCoveredEntries() function to perform the join. In this function, when
we copy the item value corresponding to the left-hand-side of the join, we always use
CopyWithCovers() function which copies the value as well as covering information. The
problem is if the left-hand-side of the index join is not using a covering index scan,
then the cover filed is nil and CopyWithCovers() will trigger a runtime panic.

The fix is to move the logic for checking and copying a non-nil covers info to the
Copy() function itself, and remove CopyWithCovers() function. The callers should
then just use the Copy() function for all situations.

Change-Id: If21b29e8bc566d84e3bc14b73705a83331ced777
Reviewed-on: http://review.couchbase.org/86591
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>