MB-27188, MB-27227 UPDATE and DELETE statements not using appropriate index
This is a regression from fix for MB-27109.
Previously the handling of WHERE clause (break down into individual filters
and attach to appropriate keyspaces) is done in function buildScan(), which
is called from two places, one for SELECT and one for UPDATE/DELETE. The fix
for MB-27109 moved the handling of WHERE clause to the root of the SELECT
statement, as a result for UPDATE/DELETE statement the appropriate processing
is missing and thus index selection does not work properly.

The fix is to also call similar functionality in UPDATE/DELETE statements.
To make it work better a couple of accessory functions are introduced, such
that we can call the accessory functions in different places.

Change-Id: I16a58e104d7735bb94649ccfdcdd4a792d4b9551
Reviewed-on: http://review.couchbase.org/86841
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>