MB-27230 incorrect error message returned for ANSI JOIN with OR clauses
In ANSI JOIN, it's possible to have an OR clause as join predicate where
one subterm of the OR does not contain any references to the keyspace
being joined. For example, b1 JOIN b2 on "b1.c1 = b2.c2 OR b1.c3 > 5",
in this case the second arm of the OR clause only contains reference to
b1, and since we are doing index selection on b2 (right-hand-side), we
cannot use any index on b2 to satisfy the OR clause. In this case, an
error (no index for ANSI JOIN term b2) should be returned. However,
currently an internal error is being returned.

Fix, check for join situation and returns a nil return scan instead of
internal error. The caller will detect the nil return and generates the
appropriate error message.

Change-Id: I41bdc8caef4fe199608add4aaa13fd5370581a74
Reviewed-on: http://review.couchbase.org/86848
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>