MB-26266 serialize certain low cost operators
The bulk of the feature is base.go.
Essentially now the sendItemOp() method uses a functional pointer to determine how to send an item.
For most cases, this is preset to parallelSend() which sends the item down value exchange channels, the usual way.

Where operators are marked as serializable, this can be set to serializedSend(), which instead inlines the consumer in the producer, which executes the consumer processItem() method directly.
To make things simpler, currently only the sequence operator will turn on inlining for consumers - which I believe is in the spirit of running operators inline.
Serializable operators will operate correctly with unchanged code both serialized and in parallel.

There is logic in runConsumer(), SetOutput() and close() to setup and propagate inlining, and take those actions that are needed to notify operators that an inline operator has stopped.

The final projection operator is safe to inline because it operates on a value and nothing else.
The offset and limit operators should use atomic operations for the offset and limit counters, but since they get fed directly from a sequence taking values from project_final (which never runs in parallel) they are safe to be run non atomically. 

  

Change-Id: I60c91151fb5df184d6b11ffb946310085eb11e09
Reviewed-on: http://review.couchbase.org/87018
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>