MB-27356 retrieve prepared statement remotely if not found locally
With this change, the prepared statement cache will retrieve a prepared statement from a remote node if the requested statement name is host qualified and the statement is not found.
Behaviour is unchanged if the prepared name is not qualified.
When I implement MB-27386 (PREPARE adds host name to statement name), the SDKs will automatically take advantage of this feature.
Added options flags for GetPrepared(), so that its behaviour can be changed.
Will later add the OPT_VERIFY flag when I implement auto_reprepare.

Change-Id: I7524c3f4b3a6ab094d80a85b4cc4ae61efa93438
Reviewed-on: http://review.couchbase.org/87326
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>