MB-27386 prepare includes node name in prepared name
Prepare now outputs the following:
cbq> prepare a from select * from `travel-sample`;
{
    "requestID": "154014de-fb9d-44c9-9352-5597594df344",
    "signature": "json",
    "results": [
        {
            "encoded_plan": "...",
            "name": "[127.0.0.1:8091]a",
            "operator": {
                "#operator": "Sequence",
            ...

Note how the name includes the node name where the statement is created.
So when the SDKs get the fully qualified name and use it as part of the "prepareds" query REST parameter, the N1QL service is able to use it as part of MB-27356 and retrieve the plan remotely if it is not found locally.
The name in the cache is not changed. Only the output of PREPARED. 

Change-Id: I5e7f4f78fe03dba029f41e2eeb68203397276468
Reviewed-on: http://review.couchbase.org/87328
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>