MB-26582 : Close and Restart TLS listener gracefully.
The TCPListener Close() does not guarantee that the address is safe to 
listen on again. This is because if a concurrent goroutine is in the 
middle of an Accept() call on the same listener, it will hold an 
additional reference, via an fdmutex readLock, to the underlying file 
descriptor. (as per the golang documentation).

On looking further into the logs we see that the following occurs
   -  In doSSLcert(), we close the tls listener and the restart it. 
      Here it fails and gives the error Bind address is already in use. 
      (This happes 2ce before the listener is actually closed)
   -  Now when we try to listen on 18093, since it was closed earlier 
      but didn’t restart correctly we get the err  - Use of closed 
      network connection.

To ensure that we dont indefinitely try to close a closed listener. 
Possible scenarios -
1. Close successful, restarting listener successful
2. Close successful, but restarting listener errors as close not 
   complete. here we sleep, and on next retry listener is restarted.
3. Close unsuccesful, so listener restarting unsuccessful - > sleep, 
   same as case 2.

Change-Id: I98e72c1ca5aa0d8426d41773887e7d75570b4d42
Reviewed-on: http://review.couchbase.org/87709
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Isha Kandaswamy <isha@couchbase.com>