MB-27500 Fix hostname to not add extra brackets [[::1]]
The net.JoinHostPort() method adds square brackets around the host name
in any input ipv6 url. However when we enter the url in cbq it already
includes the square brackets. Because of this we get nested brackets -
http://[ipv6_address]:8091 -> http://[[ipv6_address]]:8091 .

To avoid this we use the url.Hostname() instead of the url.Host to
return the raw host name.

Change-Id: Ib7af77964ac50d23465f8cd90064046e7b396adb
Reviewed-on: http://review.couchbase.org/87724
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Isha Kandaswamy <isha@couchbase.com>