MB-27357 add plan operator verification infrastructure
This is longwinded but the changes are mechanical.
Planner and server just add a little bit of extra information to prepared statement creation and plan operator creation (more later).
Datastore adds Indexer and Namespace metadata counter support.
Plan adds a verify method which checks indexes, keyspaces and children as required, and returns true if all good, false if not, and amends prepared statement metadata counters if a prepared statement is passed (this is necessary because the planner and plan infrastructure does not cater for prepared statements).
The counters will be used in MB-27358 to avoid in dept checking if the indexers and the namespaces don't report metadata changes.

Things still left to do:
1) the datastore Index interface needs to add an Indexer() method, so that, when we have an index, we can do stuff like refreshing the indexer, or saving the indexer in the prepared statement structure. This has been created for all indexers bar GSI, and I have already discussed the issue with John and Jeelan.
Some through the hoop jumping code in use right now will disappear once we have that facility.

2) MB-27358: use the verify() method from an ad hoc prepared statement Verify() method to be called whenever we get a prepared REST API parameter (service_request.go) or EXECUTE a statement (server.go).
This will check metadata counters with a read lock if present, and verify and repopulate the whole plan under a lock if the counters are missing or out of date, and finally reprepare the statement if the plan is not usable.
Initial counter population will be done at the first execution - this will check and populate in one go which saves us one check. 

Change-Id: I9119d275ccbb2f6249c46cc84c75c98af2f55588
Reviewed-on: http://review.couchbase.org/87895
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Isha Kandaswamy <isha@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>