MB-26740 Allow index keys or variables inside index keys to use same identifier as keyspace alias
There are a number of defects where an index key or a variable inside an index
key collide with a keyspace alias, and an error or potentially wrong results
can occur. Inside an index definition, an identifier should not match to
a keyspace alias since the nature of index keys cannot be a keyspace reference.

The fix is to introduce mechanism to indicate whether an identifier is a
keyspace alias, and allow collision to occur with keyspace alias when
formalizing index keys as part of index matching process (as indicated by
the mapSelf property of the formalizer). The only exception is if the index
definition uses a binding scope, for example a variable for an ANY clause,
then the expression such as the SATISFIES clause can have identifiers that
matches a keyspace alias.

Change-Id: I0f9ba2729d89e475f82941bf288cea391c814ac8
Reviewed-on: http://review.couchbase.org/87933
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>