MB_27706 Token index fields may not be formalized properly when query executed multiple times
This is a regression from fix for MB-26740.

The fix for MB-26740 added mechanism for identifing a keyspace alias as allowed
identifier, and when formalizing index keys or index conditions, allow variables
or key fields to use the same alias as a keyspace alias. The problem is that
there may be multiple places where exceptions are required, such as a binding
scope, or in case of this defect, in a tokens index. The previous checkin only
handles binding scope.

The fix is to modify the mechansim such that we now explicitly check for index
keys and index conditions, such that formalization in other area of the code
are not to be affected by the original change.

Change-Id: I7d9189459948483845e5d22e381ee94462652cba
Reviewed-on: http://review.couchbase.org/88387
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>