MB-27358 verify plans and prepare again if necessary
The bulk of the changes are splitting plan/prepared.go in two, given that it did two things at once: represent the plan, and implement a cache of prepared statements.
This gave rise to loads of circular references when adding the verify and reprepare logic.

We now have plan/prepared.go which represents the plan and prepareds/prepareds.go which contains the cache operations inclusive of remote distribution and verification and reprepare if necessary.

Getting the prepared statement for the EXECUTE statement was split between planner and server (for reasons unknown). This was messy and a duplication of work, so it was all moved to server, whit planner's VisitExecute() retained just in case we need it for the planned EXECUTE improvements.
All it does is it checks that the prepares statement value is the correct type.
It returns a discard operator because we can't return a nil interface. It's never used anyway.

Currently reprepare entails reprising the statement. This is not really efficient, but saves memory, as to skip that we would have to save the AST in the prepared statement.
Since reprepare will happen very rarely, I've gone for cpu cost rather than memory.

There are a couple more entries in the prepare output (and encoded plan): index api version and feature controls. These are needed because reprepare has to produce a plan under the same conditions as it was originally prepared.
These are included in the encoded_plan - which is currently screwing some of the indexga tests, because they check encoded plans and now these have changed, but I have not yet worked out what I need to fix. I may need Sitaram's help.

There might be a slight race condition issue that I have to fix in GetPrepared.

The encoded_plan rest api mechanism is still there and has been augmented with auto-reprepare, so that we don't plant stale plans.
Encoded_plan rest api will not trigger remote distribution, because if client applications are using it, multiple clients will and if we distribute remotely, this will probably cause a stampede with all nodes executing a request trying to distribute to remote nodes.


Change-Id: Iaf5186c4013731b34eed5366fe14e3fbd8d12f95
Reviewed-on: http://review.couchbase.org/88427
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>