MB-25757 Change plan and execution operator from AnsiJoin to NLJoin
Currently we only support index nested-loop join for ANSI JOIN at runtime,
thus the plan and execution operators are named AnsiJoin. However when we
eventually add hash join support, it may be confusing since ANSI JOIN is
the semantics name (join with ON clause) while nested-loop join and hash
join are join implementations. To avoid future confusion we are changing
the plan and execution operator names from AnsiJoin/AnsiNest to
NLJoin/NLNest. Externally, e.g. in explain, we are going to print out the
full name "NestedLoopJoin"/"NestedLoopNest". Note that in algebra the node
name is unchanged, it remains as AnsiJoin/AnsiNest, since that corresponds
to the syntax, not join methods.

Change-Id: I8256491b0ebc8adf157e4efa13598b8f34199128
Reviewed-on: http://review.couchbase.org/88518
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>