MB-25757 Change explain indicator from ansi_join to nested_loop
Currently in explain of an ANSI JOIN query, the inner side operators of ANSI JOIN
can be marked as "ansi_join": true, this is done for index scans and fetches.
This marking is important to properly evaluate spans for index scans at runtime.
However, this is really for nested-loop joins. To prepare for hash join support
change the indicator from "ansi_join": true to "nested_loop": true.

Change-Id: I4e5236a841a65a6b26b85359c91e16c09fc274a0
Reviewed-on: http://review.couchbase.org/88581
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>